refactgor:클래스들을 model 파일로 별도 분리
